move ocr tabs to sub tabs
