+++ src/components/document/TabPanel.js
@@ -1,18 +1,35 @@
 import React, { memo } from 'react'
 import { Box } from '@material-ui/core'
+import { makeStyles } from '@material-ui/core/styles'
 
-const TabPanel = ({ children, value, index, padding = 2, alwaysVisible = false, ...other }) => (
-    <div
-        role="tabpanel"
-        hidden={!alwaysVisible && value !== index}
-        {...other}
-    >
-        {(alwaysVisible || value === index) && (
-            <Box p={padding}>
-                {children}
-            </Box>
-        )}
-    </div>
-)
+const useStyles = makeStyles(theme => ({
+    root: {
+        height: '100%',
+    },
+    tab: {
+        display: 'flex',
+        flexDirection: 'column',
+        height: '100%',
+    },
+}))
+
+const TabPanel = ({ children, value, index, padding = 2, alwaysVisible = false, tabClass, ...other }) => {
+    const classes = useStyles()
+
+    return (
+        <div
+            role="tabpanel"
+            hidden={!alwaysVisible && value !== index}
+            className={classes.root}
+            {...other}
+        >
+            {(alwaysVisible || value === index) && (
+                <Box p={padding} className={classes.tab}>
+                    {children}
+                </Box>
+            )}
+        </div>
+    )
+}
 
 export default memo(TabPanel)
